+++ packages/frontend/src/widgets/documents-side-menu/index.ts
@@ -1,0 +1,3 @@
+export * from './ui/HeadMenu/HeadMenu'
+export * from './ui/SideMenu/SideMenu'
+export * from './lib/type/IPropsMenu'
