+++ packages/frontend/src/widgets/documents-side-menu/lib/type/IPropsMenu.tsx
