+++ packages/frontend/src/widgets/documents-side-menu/ui/HeadMenu/HeadMenu.module.scss
