+++ packages/frontend/src/widgets/documents-side-menu/ui/HeadMenu/HeadMenuItems.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link'
-import { IPropsMenu } from '../type/IPropsMenu'
-import style from './HeadBar.module.scss'
+import { IPropsMenu } from '../../lib/type/IPropsMenu'
+import style from './HeadMenu.module.scss'
 
-const HeadBarItem = ({ data }: {data: IPropsMenu }): JSX.Element => {
+export const HeadMenuItem = ({ data }: {data: IPropsMenu }): JSX.Element => {
   return (
     <li className={ style['menu'] }>
       <Link href={ data.href }>
@@ -14,5 +14,3 @@ const HeadBarItem = ({ data }: {data: IPropsMenu }): JSX.Element => {
 
   )
 }
-
-export default HeadBarItem
