+++ packages/frontend/src/widgets/documents-side-menu/ui/SideMenu/SideMenu.module.scss
@@ -18,6 +18,6 @@
   display: flex;
   justify-content: center;
   height: 1018px;
-  padding-top: 52px;
+  padding: 52px 40px 0 0;
   border-right: 2px solid $neutral-50;
 }
